+++ molo/forms/views.py
@@ -7,6 +7,7 @@ from rest_framework.response import Response
 from wagtail.core.models import Page
 
 from django.conf.urls import url
+from django.http.response import HttpResponse
 from django.views.generic import TemplateView, View
 from molo.forms.models import MoloFormPage, FormsIndexPage, PersonalisableForm
 from molo.core.models import ArticlePage
@@ -130,6 +131,7 @@ class ResultsPercentagesJson(View):
 
 
 class FormSuccess(TemplateView):
+    is_ajax = False
     template_name = "forms/molo_form_page_success.html"
 
     def get_context_data(self, *args, **kwargs):
@@ -141,16 +143,24 @@ class FormSuccess(TemplateView):
         form = get_object_or_404(
             MoloFormPage, slug=kwargs['slug'], id__in=ids)
         results = dict()
+
         if form.show_results:
             # Get information about form fields
             data_fields = [
                 (field.clean_name, field.label)
                 for field in form.get_form_fields()
             ]
-
             # Get all submissions for current page
+            article = kwargs.get('article')
+            if article:
+                f = {'article_page_id': article}
+            else:
+                f = {'article_page__isnull': True}
+
             submissions = (
-                form.get_submission_class().objects.filter(page=form))
+                form.get_submission_class().objects.filter(
+                    page=form, **f))
+
             for submission in submissions:
                 data = submission.get_data()
 
@@ -171,13 +181,22 @@ class FormSuccess(TemplateView):
                     question_stats = results.get(label, {})
                     question_stats[answer] = question_stats.get(answer, 0) + 1
                     results[label] = question_stats
-        if form.show_results_as_percentage:
-            for question, answers in results.items():
-                total = sum(answers.values())
-                for key in answers.keys():
-                    answers[key] = int((answers[key] * 100) / total)
+
+            if form.show_results_as_percentage:
+                for question, answers in results.items():
+                    total = sum(answers.values())
+                    for key in answers.keys():
+                        answers[key] = int((answers[key] * 100) / total)
         context.update({'self': form, 'results': results})
         return context
+
+    def dispatch(self, request, *args, **kwargs):
+        if self.is_ajax or request.GET.get('format') == 'json':
+            context = self.get_context_data(*args, **kwargs)
+            content = json.dumps(context.get('results'))
+            return HttpResponse(
+                content=content, content_type='application/json')
+        return super(FormSuccess, self).dispatch(request, *args, **kwargs)
 
 
 def submission_article(request, form_id, submission_id):